# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformAdminUser {
  id: string;
  email: string;
  name: string;
  role: string;
}

export interface PlatformAdminAuthUserResponse {
  user: PlatformAdminUser | null;
  admin: boolean;
}

export type PlatformAdminOverviewCommunication = { [key: string]: unknown };

export interface PlatformMqttConfig {
  brokerUrl: string;
  topic: string;
  plantSite: string;
  timezone: string;
  credentialsConfigured: boolean;
  pendingApply: boolean;
}

export type PlatformAuditEventMetadata = { [key: string]: unknown };

export interface PlatformAuditEvent {
  id: string;
  actorEmail: string;
  action: string;
  targetType: string;
  targetId: string;
  createdAt: string;
  metadata: PlatformAuditEventMetadata;
}

export interface PlatformAdminOverview {
  organizationCount: number;
  siteCount: number;
  userCount: number;
  activeAccessGrantCount: number;
  broker: PlatformMqttConfig;
  communication: PlatformAdminOverviewCommunication;
  recentAudit: PlatformAuditEvent[];
}

export interface PlatformOrganization {
  id: string;
  name: string;
  slug: string;
  status: string;
  siteCount: number;
}

export interface PlatformOrganizationInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  name: string;
  /**
     * @minLength 2
     * @maxLength 80
     */
  slug: string;
}

export interface PlatformSite {
  siteName: string;
  organizationId: string;
  organizationName: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  timezone: string;
  status: string;
}

export interface PlatformSiteInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  siteName: string;
  organizationId: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @maxLength 80 */
  timezone: string;
}

export interface PlatformSiteAccess {
  userId: string;
  siteName: string;
  organizationId: string;
  role: string;
  status: string;
}

export interface PlatformUser {
  id: string;
  /** @nullable */
  email: string | null;
  name: string;
  access: PlatformSiteAccess[];
}

export type PlatformSiteAccessInputRole = typeof PlatformSiteAccessInputRole[keyof typeof PlatformSiteAccessInputRole];


export const PlatformSiteAccessInputRole = {
  viewer: 'viewer',
  operator: 'operator',
  'site-admin': 'site-admin',
} as const;

export interface PlatformSiteAccessInput {
  userId: string;
  siteName: string;
  role: PlatformSiteAccessInputRole;
}

export interface PlatformMqttConfigInput {
  /**
     * @minLength 8
     * @maxLength 300
     */
  brokerUrl: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  topic: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  plantSite: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  timezone: string;
}