# lib/api-zod/src/generated/types/platformAdminOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformAdminOverviewCommunication } from './platformAdminOverviewCommunication';
import type { PlatformAuditEvent } from './platformAuditEvent';
import type { PlatformMqttConfig } from './platformMqttConfig';

export interface PlatformAdminOverview {
  organizationCount: number;
  siteCount: number;
  userCount: number;
  activeAccessGrantCount: number;
  broker: PlatformMqttConfig;
  communication: PlatformAdminOverviewCommunication;
  recentAudit: PlatformAuditEvent[];
}