# lib/api-zod/src/generated/types/platformMqttConfigInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformMqttConfigInput {
  /**
     * @minLength 8
     * @maxLength 300
     */
  brokerUrl: string;
  /**
     * @minLength 1
     * @maxLength 300
     */
  topic: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  plantSite: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  timezone: string;
}