# lib/api-zod/src/generated/types/platformOrganizationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformOrganizationInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  name: string;
  /**
     * @minLength 2
     * @maxLength 80
     */
  slug: string;
}