# lib/api-zod/src/generated/types/platformSite.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformSite {
  siteName: string;
  organizationId: string;
  organizationName: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  timezone: string;
  status: string;
}