# lib/api-zod/src/generated/types/platformSiteAccess.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformSiteAccess {
  userId: string;
  siteName: string;
  organizationId: string;
  role: string;
  status: string;
}