# lib/api-zod/src/generated/types/platformSiteAccessInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformSiteAccessInputRole } from './platformSiteAccessInputRole';

export interface PlatformSiteAccessInput {
  userId: string;
  siteName: string;
  role: PlatformSiteAccessInputRole;
}