# lib/api-zod/src/generated/types/platformSiteInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformSiteInput {
  /**
     * @minLength 2
     * @maxLength 160
     */
  siteName: string;
  organizationId: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @maxLength 80 */
  timezone: string;
}