# lib/api-zod/src/generated/types/platformUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlatformSiteAccess } from './platformSiteAccess';

export interface PlatformUser {
  id: string;
  /** @nullable */
  email: string | null;
  name: string;
  access: PlatformSiteAccess[];
}